fixed manual template matmul MNK
