+++ autotune/modules/dma.py
@@ -46,6 +46,7 @@ def load_tensor_block(input_tensor, ofs: Tuple[int, int], load_shape: Tuple[int,
         for column_tile_id in nl.affine_range(column_num_tiles):
             row_indices = ofs[0] + row_tile_id * row_tile_size + tile_index.p
             col_indices = ofs[1] + column_tile_id * column_tile_size + tile_index.x
+            print(f"row_indices = {row_indices}. col_indices = {col_indices}.")
             block[tile_index.p, row_tile_id, column_tile_id, tile_index.x] = nl.load(
                 input_tensor[row_indices, col_indices], mask=(row_indices < max_rows) & (col_indices < max_cols)
             )
@@ -130,6 +131,7 @@ def save_result_block(result, result_block, tile_index_ofs: Tuple[int, int]):
     """
     M, N = result.shape
     TILE_M, TILES_IN_M, TILES_IN_N, TILE_N = result_block.shape
+    print(f"Saving result_block {result_block.shape} into result {result.shape} with ofs {tile_index_ofs}.")
 
     idx_res = nl.mgrid[0:TILE_M, 0:TILE_N]
     for tile_id_M in nl.affine_range(TILES_IN_M):
